pyproject.toml
[tool.poetry]
name = "property-tax-agent"
version = "1.0.0"
description = "NZ rental property tax return document review system"
authors = ["Your Name <you@example.com>"]
readme = "README.md"
packages = [{include = "app"}]

[tool.poetry.dependencies]
python = "^3.12"
fastapi = "^0.115.5"
uvicorn = {extras = ["standard"], version = "^0.32.1"}
anthropic = "^0.75.0"
asyncpg = "^0.30.0"
sqlalchemy = "^2.0.36"
alembic = "^1.14.0"
pydantic = "^2.10.3"
pydantic-settings = "^2.6.1"
python-multipart = "^0.0.19"
pillow = "^11.0.0"
pdf2image = "^1.17.0"
pypdf = "^5.1.0"
jinja2 = "^3.1.4"
python-dotenv = "^1.0.1"
structlog = "^24.4.0"
httpx = "^0.28.1"
greenlet = "^3.3.0"
psycopg2-binary = "^2.9.11"
openpyxl = "^3.1.5"
pandas = "^2.3.3"
xlrd = "^2.0.2"
pinecone = "^8.0.0"
openai = "^2.9.0"
cryptography = "^46.0.3"
pymupdf = "^1.26.7"
pdfplumber = "^0.11.8"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.4"
pytest-asyncio = "^0.25.0"
pytest-cov = "^6.0.0"
black = "^24.10.0"
isort = "^5.13.2"
mypy = "^1.14.0"
ruff = "^0.8.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"

[tool.black]
line-length = 100
target-version = ['py312']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.12"
warn_return_any = true
warn_unused_configs = true
ignore_missing_imports = true

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
pythonpath = ["."]